Split Code::Gen() functions into their own file
